pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rpi-network-guard"
version = "1.0.0"
description = "Niskopoziomowy, asynchroniczny agent monitorujący sieć dedykowany dla systemów o ograniczonych zasobach (Raspberry Pi)."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "Dominik (IT & CyberSec)", email = "dominik@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
